+++ changelog.md
@@ -25,7 +25,7 @@
 
 - minimal supported version of *Angular* is `10.0.0`
 - minimal supported version of `@jscrpt/common` is `1.2.0`
-- removed `HttpRequestIgnoredInterceptorId` interface, explicitly using type unions where needed
+- removed `HttpRequestIgnoredInterceptorId` interface, explicitly using type unions where needed, `HttpRequest<any> & AdditionalInfo<IgnoredInterceptorId>`
 - changed `IgnoredInterceptorId` to generic `AdditionalInfo`, allowing broader usage
 
 ## Version 7.3.2 (2020-05-15)